+++ besu/src/main/java/org/hyperledger/besu/controller/CliqueBesuControllerBuilder.java
@@ -130,7 +130,9 @@ public class CliqueBesuControllerBuilder extends BesuControllerBuilder {
 
   @Override
   protected CliqueContext createConsensusContext(
-      final Blockchain blockchain, final WorldStateArchive worldStateArchive) {
+      final Blockchain blockchain,
+      final WorldStateArchive worldStateArchive,
+      final ProtocolSchedule protocolSchedule) {
     return new CliqueContext(
         BlockValidatorProvider.nonForkingValidatorProvider(
             blockchain, epochManager, blockInterface),