+++ besu/src/main/java/org/hyperledger/besu/controller/QbftBesuControllerBuilder.java
@@ -14,10 +14,10 @@
  */
 package org.hyperledger.besu.controller;
 
-import org.hyperledger.besu.config.BftConfigOptions;
 import org.hyperledger.besu.config.BftFork;
 import org.hyperledger.besu.config.GenesisConfigOptions;
 import org.hyperledger.besu.config.QbftConfigOptions;
+import org.hyperledger.besu.consensus.common.BftValidatorOverrides;
 import org.hyperledger.besu.consensus.common.EpochManager;
 import org.hyperledger.besu.consensus.common.bft.BftContext;
 import org.hyperledger.besu.consensus.common.bft.BftEventQueue;
@@ -46,11 +46,15 @@ import org.hyperledger.besu.consensus.qbft.QbftExtraDataCodec;
 import org.hyperledger.besu.consensus.qbft.QbftGossip;
 import org.hyperledger.besu.consensus.qbft.jsonrpc.QbftJsonRpcMethods;
 import org.hyperledger.besu.consensus.qbft.payload.MessageFactory;
+import org.hyperledger.besu.consensus.qbft.pki.PkiQbftContext;
+import org.hyperledger.besu.consensus.qbft.pki.PkiQbftExtraDataCodec;
 import org.hyperledger.besu.consensus.qbft.protocol.Istanbul100SubProtocol;
 import org.hyperledger.besu.consensus.qbft.statemachine.QbftBlockHeightManagerFactory;
 import org.hyperledger.besu.consensus.qbft.statemachine.QbftController;
 import org.hyperledger.besu.consensus.qbft.statemachine.QbftRoundFactory;
 import org.hyperledger.besu.consensus.qbft.validation.MessageValidatorFactory;
+import org.hyperledger.besu.consensus.qbft.validator.TransactionValidatorProvider;
+import org.hyperledger.besu.consensus.qbft.validator.ValidatorContractController;
 import org.hyperledger.besu.ethereum.ProtocolContext;
 import org.hyperledger.besu.ethereum.api.jsonrpc.methods.JsonRpcMethods;
 import org.hyperledger.besu.ethereum.blockcreation.MiningCoordinator;
@@ -67,6 +71,7 @@ import org.hyperledger.besu.ethereum.eth.sync.state.SyncState;
 import org.hyperledger.besu.ethereum.eth.transactions.TransactionPool;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
 import org.hyperledger.besu.ethereum.p2p.config.SubProtocolConfiguration;
+import org.hyperledger.besu.ethereum.transaction.TransactionSimulator;
 import org.hyperledger.besu.ethereum.worldstate.WorldStateArchive;
 import org.hyperledger.besu.util.Subscribers;
 
@@ -89,7 +94,14 @@ public class QbftBesuControllerBuilder extends BftBesuControllerBuilder {
 
   @Override
   protected Supplier<BftExtraDataCodec> bftExtraDataCodec() {
-    return Suppliers.memoize(QbftExtraDataCodec::new);
+    return Suppliers.memoize(
+        () -> {
+          if (pkiBlockCreationConfiguration.isPresent()) {
+            return new PkiQbftExtraDataCodec();
+          } else {
+            return new QbftExtraDataCodec();
+          }
+        });
   }
 
   @Override
@@ -258,23 +270,44 @@ public class QbftBesuControllerBuilder extends BftBesuControllerBuilder {
 
   @Override
   protected BftContext createConsensusContext(
-      final Blockchain blockchain, final WorldStateArchive worldStateArchive) {
+      final Blockchain blockchain,
+      final WorldStateArchive worldStateArchive,
+      final ProtocolSchedule protocolSchedule) {
     final GenesisConfigOptions configOptions =
         genesisConfig.getConfigOptions(genesisConfigOverrides);
-    final BftConfigOptions bftConfig = configOptions.getBftConfigOptions();
-    final EpochManager epochManager = new EpochManager(bftConfig.getEpochLength());
+    final QbftConfigOptions qbftConfig = configOptions.getQbftConfigOptions();
+    final EpochManager epochManager = new EpochManager(qbftConfig.getEpochLength());
 
-    final Map<Long, List<Address>> bftValidatorForkMap =
+    final BftValidatorOverrides validatorOverrides =
         convertBftForks(configOptions.getTransitions().getQbftForks());
 
-    return new BftContext(
-        BlockValidatorProvider.forkingValidatorProvider(
-            blockchain, epochManager, bftBlockInterface().get(), bftValidatorForkMap),
-        epochManager,
-        bftBlockInterface().get());
+    final ValidatorProvider validatorProvider;
+    if (qbftConfig.getValidatorContractAddress().isEmpty()) {
+      validatorProvider =
+          BlockValidatorProvider.forkingValidatorProvider(
+              blockchain, epochManager, bftBlockInterface().get(), validatorOverrides);
+    } else {
+      final Address contractAddress =
+          Address.fromHexString(qbftConfig.getValidatorContractAddress().get());
+      final TransactionSimulator transactionSimulator =
+          new TransactionSimulator(blockchain, worldStateArchive, protocolSchedule);
+      final ValidatorContractController validatorContractController =
+          new ValidatorContractController(contractAddress, transactionSimulator);
+      validatorProvider = new TransactionValidatorProvider(blockchain, validatorContractController);
+    }
+
+    if (pkiBlockCreationConfiguration.isPresent()) {
+      return new PkiQbftContext(
+          validatorProvider,
+          epochManager,
+          bftBlockInterface().get(),
+          pkiBlockCreationConfiguration.get());
+    } else {
+      return new BftContext(validatorProvider, epochManager, bftBlockInterface().get());
+    }
   }
 
-  private Map<Long, List<Address>> convertBftForks(final List<BftFork> bftForks) {
+  private BftValidatorOverrides convertBftForks(final List<BftFork> bftForks) {
     final Map<Long, List<Address>> result = new HashMap<>();
 
     for (final BftFork fork : bftForks) {
@@ -288,7 +321,7 @@ public class QbftBesuControllerBuilder extends BftBesuControllerBuilder {
                           .collect(Collectors.toList())));
     }
 
-    return result;
+    return new BftValidatorOverrides(result);
   }
 
   private static MinedBlockObserver blockLogger(