+++ ethereum/retesteth/src/main/java/org/hyperledger/besu/ethereum/retesteth/methods/TestSetChainParams.java
@@ -14,7 +14,6 @@
  */
 package org.hyperledger.besu.ethereum.retesteth.methods;
 
-import org.hyperledger.besu.config.experimental.ExperimentalEIPs;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.JsonRpcRequestContext;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.JsonRpcMethod;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError;
@@ -133,16 +132,7 @@ public class TestSetChainParams implements JsonRpcMethod {
     maybeMove(genesis, "nonce", chainParamsJson, "nonce");
     maybeMove(genesis, "timestamp", chainParamsJson, "timestamp");
     maybeMove(chainParamsJson, "accounts", chainParamsJson, "alloc");
-
-    if (params.containsKey("londonForkBlock")) {
-      JsonObject genesisConfig = chainParamsJson.getJsonObject("genesis", new JsonObject());
-      ExperimentalEIPs.initialBasefee =
-          Optional.ofNullable(
-                  genesisConfig.getString("baseFeePerGas", genesisConfig.getString("baseFee")))
-              .map(Long::decode)
-              .orElse(ExperimentalEIPs.EIP1559_BASEFEE_DEFAULT_VALUE);
-    }
-
+    maybeMove(genesis, "baseFeePerGas", chainParamsJson, "baseFeePerGas");
     maybeMoveToNumber(params, "londonForkBlock", config, "londonBlock");
 
     // strip out precompiles with zero balance